Flask/requirements.txt
py-solc-x==2.0.2
wheel
pandas==2.1.4
numpy==1.25.2
pillow==10.2.0
hexbytes==1.0.0
pyevmasm==0.2.3
efficientnet==1.1.1
keras==2.15.0
tensorflow==2.15.0
Flask==3.0.0
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
scikit-learn==1.3.2